requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.40.0
colorama==0.4.6
cryptography==42.0.7
cssbeautifier==1.15.1
Django==4.2.13
django-ckeditor-5==0.2.13
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-filter==24.2
django-oauth-toolkit==2.4.0
djangorestframework==3.15.1
drf-yasg==1.21.7
EditorConfig==0.12.4
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.7
inflection==0.5.1
jsbeautifier==1.15.1
json5==0.9.25
jwcrypto==1.5.6
Markdown==3.6
oauthlib==3.2.2
packaging==24.0
pathspec==0.12.1
pillow==10.3.0
psycopg2-binary==2.9.9
pycparser==2.22
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
regex==2024.5.15
reportlab==4.2.0
requests==2.32.2
six==1.16.0
sqlparse==0.5.0
tomli==2.0.1
toposort==1.10
tqdm==4.66.4
typing_extensions==4.11.0
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==1.26.18
